+++ servers/api/index.js
@@ -9,12 +9,15 @@ nconf
     .argv()
     .env()
     .file('general', {file: 'config/general.json'})
-    .file('deployment', {file: 'config/' + process.env.NODE_ENV + '.json'})
-    .file('common', {file: '../config/' + process.env.NODE_ENV + '.json'});
+    .file('deployment', {file: 'config/' + process.env.NODE_ENV + '.json'});
 
 var Hapi = require('hapi');
 var Joi = require('joi');
 var models = require('./models');
+var fs = require('fs');
+
+// SSL TODO:
+//  - Need to get certificate from authority.
 
 exports.init = function (testing, doneCallback) {
 
@@ -26,14 +29,54 @@ exports.init = function (testing, doneCallback) {
                     cors: {
                         origin: nconf.get('htmlOrigin'),
                         credentials: true
+                    },
+                    security: {
+                        hsts: 1000 * 60 * 60 * 24 * 365
                     }
                 }
-            }
+            },
+            cache: [
+                {
+                    name: 'redisCache', //By commenting, replace default cache
+                    engine: require('catbox-redis'),
+                    host: '127.0.0.1',
+                    partition: 'cache'
+                }
+            ]
         });
 
         server.connection({
             host: nconf.get('listenIp'),
-            port: nconf.get('port')
+            port: nconf.get('port'),
+            tls: {
+                key: fs.readFileSync(nconf.get('ssl:keyPath')),
+                cert: fs.readFileSync(nconf.get('ssl:certPath')),
+                ca: [
+                    fs.readFileSync(nconf.get('ssl:ca:1')),
+                    fs.readFileSync(nconf.get('ssl:ca:2')),
+                    fs.readFileSync(nconf.get('ssl:ca:3'))
+                ],
+                // default node 0.12 ciphers with RC4 disabled
+                ciphers: [
+                    'ECDHE-RSA-AES256-SHA384',
+                    'DHE-RSA-AES256-SHA384',
+                    'ECDHE-RSA-AES256-SHA256',
+                    'DHE-RSA-AES256-SHA256',
+                    'ECDHE-RSA-AES128-SHA256',
+                    'DHE-RSA-AES128-SHA256',
+                    'HIGH',
+                    '!aNULL',
+                    '!eNULL',
+                    '!EXPORT',
+                    '!DES',
+                    '!RC4',
+                    '!MD5',
+                    '!PSK',
+                    '!SRP',
+                    '!CAMELLIA'
+                ].join(':'),
+                honorCipherOrder: true
+            }
         });
 
         server.views({
@@ -49,6 +92,7 @@ exports.init = function (testing, doneCallback) {
         var plugins = [
             require('bell'),
             require('hapi-auth-cookie'),
+            require('hapi-auth-bearer-token'),
             {
                 register: require('hapi-swagger'),
                 options: {
@@ -68,22 +112,22 @@ exports.init = function (testing, doneCallback) {
                     reporters: [
                         {
                             reporter: require('good-console'),
-                            args: [{
+                            events: {
                                 log: '*',
                                 response: '*',
                                 error: '*'
-                            }]
+                            }
                         },
                         {
                             reporter: require('good-file'),
-                            args: [
-                                {path: nconf.get('logFilePath')},
-                                {
-                                    log: '*',
-                                    response: '*',
-                                    error: '*'
-                                }
-                            ]
+                            events: {
+                                log: '*',
+                                response: '*',
+                                error: '*'
+                            },
+                            config: {
+                                path: nconf.get('logFilePath')
+                            }
                         }
                     ]
                 }